requirements.txt
insightface
opencv-python
numpy
scipy
Pillow
onnxruntime
fastapi
uvicorn[standard]
python-multipart